Introduces "repeat last run" feature
Sometime, editing the spec file after a failed test, you want to rerun
the last example, but being outside you had to go back to place the
cursor inside that test. With this change it's a matter of another
keystroke to repeat the last run regardless of the cursor position.