Fixes handling the errors-outside-of-examples case
RSpec reports "Failure/Error" as a plain text and as part of the json
output. The former is a general case handled by the runner and the
latter should be passed to the file integration.

Changing the match pattern so that "Failure/Error" is not part of
a JSON string.